requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
motor==3.3.1
pydantic==2.3.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.0.1
qrcode==7.4.2
pillow==10.0.0
pytest==7.4.2
httpx==0.25.0
